feat: apply lazy load for rest-client page
