+++ src/shared/i18n/messages/ru.json
@@ -198,6 +198,7 @@
   },
   "History": {
     "title": "История и Аналитика",
+    "loading": "Загрузка...",
     "emptyTitle": "Вы ещё не выполняли запросов. ",
     "emptySubtitle": "Здесь пусто. Попробуйте сделать запрос",
     "goRestClient": "Перейти в REST клиент",
